# Speed up v3 service plans by removing eager loading of visibilities