--- a/lib/tasks/db.rake
+++ b/lib/tasks/db.rake
@@ -121,6 +121,16 @@ namespace :db do
     end
   end
 
+  desc 'Seed the database with a lot of data for performance testing'
+  task :seed_perf_data do
+    RakeConfig.context = :api
+
+    require 'cloud_controller/perf_data_seeds'
+    BackgroundJobEnvironment.new(RakeConfig.config).setup_environment do
+      VCAP::CloudController::PerfDataSeeds.write_seed_data(RakeConfig.config)
+    end
+  end
+
   desc 'Migrate and seed database'
   task :setup_database do
     Rake::Task['db:migrate'].invoke