## Overview
**A short explanation of the proposed change:**
Remove eager loading of service plan visibilities when fetching the list of service plans

> Note: there are 2 commits here - the 1st one just adds a helpful rake task for populating the DB and isn't critical to the behaviour, we just included it to help test locally/test in the future. If whoever reviews this would rather just take the functional second commit please let us know

**An explanation of the use cases your change solves**
Users trying to query `cf marketplace` against deployments with large numbers of organisations and a reasonably large number of brokers/services/plans which thus have an enormous number of service plan visibilities

**Links to any other associated PRs**
Fixes #2213 

## More info
After a bunch of investigation we found that the service plan visibilities table was causing issues when eagerly loaded into the service plans endpoint, particularly when listing service plans (as `cf marketplace` does). Whilst the query that the ORM executes (`SELECT * FROM service_plan_visibilities WHERE service_plan_id IN (<long list of IDs>)`) isn't that expensive for a DB, it returns >10k rows in many cases and the process of parsing that to Ruby objects and linking it to the correct plan does seem to be pretty slow.

The only usage of this data in this endpoint appears to be [checking whether there are any visibilities](https://github.com/cloudfoundry/cloud_controller_ng/blob/9c796d83da5723d576e8430fa30186c063e6fdcb/app/models/services/service_plan.rb#L167-L175) for this plan and only in the case where the plan is not public or not space-scoped. By _not_ eagerly loading this table, the ORM instead issues a `SELECT 1 ... WHERE service_plan_id = '1234'` for each plan in the list but as this a) is only called if it is not public or space scoped and b) returns much less data it is far faster than the eager loading still. _Potentially_, this may remain true in environments with **significant** DB latency as more, smaller requests are needed but as the numbers below will say it would have to be **really** bad to be worse than it currently is.

## Tests
Setting up a database as per [this comment](https://github.com/cloudfoundry/cloud_controller_ng/issues/1591#issuecomment-733059916):
* 22,000 orgs
* 66,000 spaces
* 300 service brokers
* 600 services
* 1200 service plans
* 660,000 service plan visibilities (all plans have worst-case scenario of being organization visible)

We got the following timings for querying the `service_plan` endpoints before and after this change:
* `/v2/service_plans?results-per-page=50`  - 0.12s
* `/v2/service_plans?results-per-page=100` - 0.17s
Before change:
* `/v3/service_plans?per_page=50` - 1.24s
* `/v3/service_plans?per_page=100` - 2.21s
* `/v3/service_plans?per_page=5000` - 30.20s
After change:
* `/v3/service_plans?per_page=50` - 0.16s
* `/v3/service_plans?per_page=100` - 0.22s
* `/v3/service_plans?per_page=5000` - 1.72s

Unfortunately this hasn't actually done much for CLI users yet as there seems to be something related to the `space_guids` filter (it hasn't made it worse) but at the very least this clearly makes the basic endpoint much faster for other users (e.g. Stratos, AppsMan)

## Checklist

* [x] I have reviewed the [contributing guide](https://github.com/cloudfoundry/cloud_controller_ng/blob/main/CONTRIBUTING.md)

* [x] I have viewed, signed, and submitted the Contributor License Agreement

* [x] I have made this pull request to the `main` branch

* [x] I have run all the unit tests using `bundle exec rake`

* [ ] I have run [CF Acceptance Tests](https://github.com/cloudfoundry/cloud_controller_ng#cf-acceptance-tests-cats)