@sethboyles 2021-04-21
Interesting!  Thanks for the detailed investigation.  

If the including the n+1 ends up being slow for some cases, we could consider creating a new association on the ServicePlan model `:organization, through: :service_plan_visibilities`, which would allow us to check the same information but reduce the scope from Org count * Service Plan count to just Org count.  However, given that this query isn't needed for every ServicePlan, it still might not be worth it.