@andy-paine 2021-04-26
After doing some more profiling with `rbspy`, it looks like the majority of the time is spent parsing the `created_at` and `updated_at` columns from the `service_plan_visibilities` table. If we don't want to remove eager loading entirely I think we could probably just add a [:select](https://sequel.jeremyevans.net/rdoc/files/doc/association_basics_rdoc.html#label-3Aselect) to the visibilities association in service plans.

I tested this locally and it is not as big an improvement as removing the eager loading entirely but it is still ~6x faster I think (early numbers). That would suggest that even the worst case "I need to run 5000 `SELECT 1` queries" is better but I think it is interesting to note that most of the processing time when parsing that table is dealing with the datetimes.