@sethboyles 2021-04-29
Closing this as we cherry-picked the relevant commit: https://github.com/cloudfoundry/cloud_controller_ng/commit/2e363c69ecaa410ae713ca6a042928ef84ac05c6

We tested this out on one of our environments and everything looked good.  FWIW we did quickly try a version of my above suggestion of eager-loading a newly defined `organizations` association and did not see a performance improvement.  But I think we discovered why there was still an `n+1`, even after eager loading `service_plan_visibilities`--we believe it is because this method: 

https://github.com/cloudfoundry/cloud_controller_ng/blob/main/app/models/services/service_plan.rb#L172

calls `service_plan_visibilities_dataset` and not `service_plan_visibilities`.

as a side note there is probably a minor perf improvement to call `service_plan_visibilities_dataset.count > 0` instead of `empty?`, since `empty?` still loads the ruby models into memory and `count` executes a SQL `COUNT` statement.   

Again, thanks for this investigation and PR!